refactor: #12 패키지 구조를 도메인 별로 구별해서 리펙토링
