+++ src/main/java/com/study/sns/domain/article/model/entity/Article.java
