+++ src/main/java/com/study/sns/domain/user/account/jwt/filter/JwtTokenFilter.java
@@ -2,7 +2,7 @@ package com.study.sns.jwt.filter;
 
 import com.study.sns.jwt.UserPrincipal;
 import com.study.sns.jwt.JwtService;
-import com.study.sns.service.UserService;
+import com.study.sns.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
