+++ src/main/java/com/study/sns/domain/user/model/constant/UserRole.java
