+++ src/main/java/com/study/sns/domain/user/model/constant/UserStatus.java
